fix some error

+++ z PHP install/supervisor_php.sh
@@ -1,5 +1,7 @@
 	# install
+	if [ ! -f "/etc/supervisor/supervisord.conf" ]; then
 	apt-get install -y --no-install-recommends supervisor
+	fi
 	# Supervisor config
 		[[ -d /var/log/supervisor ]] || mkdir -p /var/log/supervisor/
 		[[ -d /etc/supervisor/conf.d ]] || mkdir -p /etc/supervisor/conf.d/
